+++ octave/expRegression.m
@@ -1,4 +1,0 @@
-function [Y P] = expRegression(X)
-  [Y P] = linearRegression(log(X));
-  Y = exp(Y);
-endfunction