+++ octave/logRegression.m
@@ -1,4 +1,0 @@
-function [Y P] = logRegression(X)
-  [Y P] = linearRegression(exp(X));
-  Y = log(Y);
-endfunction