+++ src/main/java/org/mmarini/wheelly/apps/Yaml.java
@@ -39,6 +39,12 @@ import static org.mmarini.yaml.schema.Validator.*;
 
 public interface Yaml {
 
+    Validator BASE_CONFIG = objectPropertiesRequired(Map.of(
+            "version", string(values("0.2")),
+            "active", string(),
+            "configurations", object()
+    ), List.of("version", "active", "configurations"));
+
     static Validator analysis() {
         return objectPropertiesRequired(Map.of(
                         "version", string(values("0.1")),
@@ -56,14 +62,6 @@ public interface Yaml {
                         "numEpochs"
                 )
         );
-    }
-
-    static Validator baseConfig() {
-        return objectPropertiesRequired(Map.of(
-                "version", string(values("0.2")),
-                "active", string(),
-                "configurations", object()
-        ), List.of("version", "active", "configurations"));
     }
 
     static Validator createFile() {