+++ src/test/java/org/mmarini/wheelly/agents/TDAgentActObsTest.java
@@ -114,7 +114,9 @@ class TDAgentActObsTest {
         JsonNode criticSpec = Utils.fromText(CRITIC_YAML);
         TDNetwork critic = TDNetwork.create(criticSpec, Locator.root(), "", Map.of(), random);
         return new TDAgent(STATE_SPEC, ACTIONS_SPEC,
-                0, REWARD_ALPHA, ALPHA, ALPHA, LAMBDA, policy, critic, random, null, Integer.MAX_VALUE);
+                0, REWARD_ALPHA, ALPHA, ALPHA, LAMBDA,
+                policy, critic, null,
+                random, null, Integer.MAX_VALUE);
     }
 
     @Test