+++ src/test/java/org/mmarini/wheelly/agents/TDAgentTrainTest.java
@@ -93,7 +93,8 @@ class TDAgentTrainTest {
         TDNetwork critic = TDNetwork.create(criticSpec, Locator.root(), "", Map.of(), random);
         return new TDAgent(STATE_SPEC, ACTIONS_SPEC0,
                 0, REWARD_ALPHA, ALPHA, ALPHA, LAMBDA,
-                policy, critic, random, null, Integer.MAX_VALUE);
+                policy, critic, null,
+                random, null, Integer.MAX_VALUE);
     }
 
     /**